wordpress setup is reachable

+++ srcs/requirements/wordpress/tools/startup.sh
@@ -1,18 +1,17 @@
 #!/bin/bash
 
-#downloading wp cli
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
-chmod +x /wp-cli.phar
-mv wp-cli.phar /usr/local/bin/wp
+cd /var/www/html
 
-#downloading wp core
-wp core download --allow-root
+#downloading wp core when it is not done yet
+if [ ! -f wp-config.php ]; then
+	echo "downloading wordpress core files"
+	wp core download --allow-root
+	mv wp-config-sample.php wp-config.php
 
-mv wp-config-sample.php wp-config.php
-
-sed -i -e "s/database_name_here/$MARIADB_DATABASE/" \
+	sed -i -e "s/database_name_here/$MARIADB_DATABASE/" \
 	-e "s/username_here/$MARIADB_USER/" \
 	-e "s/password_here/$MARIADB_PASSWORD/" \
-	-e "s/localhost/mariadb:3306/" wp-config.php
+	-e "s/localhost/$MARIADB_CONNECTION/" wp-config.php
+fi
 
-sleep infinity
+php-fpm8.2 -F
